frontend/requirements.txt
flask==3.0.3
requests==2.32.4
python-dotenv==1.0.0
pandas==2.1.3
SQLAlchemy==2.0.30
